requirements.txt
feather-format==0.4.1
gender-guesser==0.4.0
numpy==1.21.2
pandas==1.3.3
pyarrow==5.0.0
python-dateutil==2.8.2
pytz==2021.3
selenium==3.141.0
six==1.16.0
urllib3==1.26.7
